requirements.txt
absl-py
flax
numpy
scipy
pandas
matplotlib
pytest
pytest-cov
seaborn
jax
jaxlib
glmnet_py
tensorflow==2.2.0rc1
tensorflow_probability
xarray
